fix: some compile warnings in MSVC

+++ test/cpp-tableau-loader/src/protoconf/scheduler.pc.cc
@@ -31,7 +31,7 @@ int Scheduler::LoopOnce() {
   for (auto&& job : jobs) {
     job();
   }
-  count += jobs.size();
+  count += static_cast<int>(jobs.size());
   return count;
 }
